% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter-generics.R
\name{getters}
\alias{getters}
\alias{year_of}
\alias{yday0_of}
\alias{yday1_of}
\alias{mon0_of}
\alias{mon1_of}
\alias{mday0_of}
\alias{mday1_of}
\alias{wday0_of}
\alias{wday1_of}
\alias{wday7_of}
\alias{n0_of_same_wday_in_year_up_to}
\alias{n1_of_same_wday_in_year_up_to}
\title{Extract "unique" information about time points or time intervals}
\usage{
year_of(obj)

yday0_of(obj)

yday1_of(obj)

mon0_of(obj)

mon1_of(obj)

mday0_of(obj)

mday1_of(obj)

wday0_of(obj)

wday1_of(obj)

wday7_of(obj)

n0_of_same_wday_in_year_up_to(obj)

n1_of_same_wday_in_year_up_to(obj)
}
\arguments{
\item{obj}{object representing a time point or time interval}
}
\description{
Here, "unique" means that that every possible time point or time interval of
an accepted class should be guaranteed to have exactly one plausible value.
This is meant to avoid bugs by forbidding, for example, \code{year_of()} from
operating on \code{\link{calweek}}s, since a single week can contain days belonging to
two different years (see \code{\link[=owning_year_of]{owning_year_of()}} instead).

These functions are all vectorized in such a way that the output should have
a \code{\link[vctrs:vec_size]{vctrs::vec_size()}} equal to that of \code{obj} if it's the only argument, or
the \code{\link[vctrs:vec_size]{vctrs::vec_size_common()}} of all arguments if there are multiple.
}
\section{Functions}{
\itemize{
\item \code{year_of()}: The uniquely-associated year, as an integer

\item \code{yday0_of()}: For the uniquely-associated date, the 0-indexed number of
days it is from the start of the containing year (e.g., January 1 of of
any year has a \code{yday0_of} of 0)

\item \code{yday1_of()}: For the uniquely-associated date, the 1-indexed number of
days it is from the start of the containing year (e.g., January 1 of any
year has a \code{yday1_of} of 1); this matches \code{\link[lubridate:day]{lubridate::yday()}}

\item \code{mon0_of()}: For the uniquely-associated month, the 0-indexed number
of months it is from the start of the containing year (e.g., January of any
year has a \code{mon0_of} of 0); this matches \verb{$mon} on \code{POSIXlt}s

\item \code{mon1_of()}: For the uniquely-associated month, the 1-indexed number
of months it is from the start of the containing year (e.g., January of any
year has a \code{mon1_of} of 1); this matches \code{\link[lubridate:month]{lubridate::month()}}

\item \code{mday0_of()}: For the uniquely-associated date, the 0-indexed number of
days it is from the start of the containing month (e.g., January 1 of any
year has a \code{mday0_of} of 0)

\item \code{mday1_of()}: For the uniquely-associated date, the 1-indexed number of
days it is from the start of the containing month (e.g., January 1 of any
year has a \code{mday1_of} of 1); this matches \verb{$mday} of \code{POSIXlt}s and
\code{lubridate::mday()}

\item \code{wday0_of()}: For the uniquely-associated date, the 0-indexed number of
days it is from the start of the containing Sunday-to-Saturday week (e.g.,
Sunday of any week has a \code{wday0_of} of 0); this matches \verb{$wday} of \code{POSIXlt}s

\item \code{wday1_of()}: For the uniquely-associated date, the 1-indexed number of
days it is from the start of the containing Sunday-to-Saturday week (e.g.,
Sunday of any week has a \code{wday1_of} of 1); this matches \code{\link[lubridate:day]{lubridate::wday()}}

\item \code{wday7_of()}: For the uniquely-associated date, the "7-indexed" number
of days it is from the start of the containing Sunday-to-Saturday week
(e.g., Sunday of any week has a \code{wday7_of} of 7, and Monday has a
\code{wday7_of} of 1); this matches "\%u" of \code{\link[=format.Date]{format.Date()}} (converted to an
integer)

\item \code{n0_of_same_wday_in_year_up_to()}: For the uniquely-associated date, the 0-indexed number of
days with the same day of week it is from the start of the containing year
(e.g., January 1 through January 6 of any year each have a
\code{n0_of_same_wday_in_year_up_to} of 0)

\item \code{n1_of_same_wday_in_year_up_to()}: For the uniquely-associated date, the 1-indexed number of
days with the same day of week it is from the start of the containing year
(e.g., January 1 through January 6 of any year each have a
\code{n1_of_same_wday_in_year_up_to } of 1)

}}
